Move `cli/remote_v1.py` under `utils`
