+++ tests/integration/conftest.py
@@ -8,7 +8,7 @@ from logging import getLogger as get_logger
 
 import pytest
 
-from milatools.cli.remote_v1 import RemoteV1
+from milatools.utils.remote_v1 import RemoteV1
 from milatools.utils.remote_v2 import (
     SSH_CONFIG_FILE,
     RemoteV2,
