+++ tests/integration/test_code_command.py
@@ -10,8 +10,8 @@ from logging import getLogger as get_logger
 import pytest
 
 from milatools.cli.commands import check_disk_quota, code
-from milatools.cli.remote_v1 import RemoteV1
 from milatools.cli.utils import get_fully_qualified_hostname_of_compute_node
+from milatools.utils.remote_v1 import RemoteV1
 from milatools.utils.remote_v2 import RemoteV2
 
 from ..cli.common import in_github_CI, skip_param_if_on_github_ci
